-Added GetSaveDir() function to Directories.h and .cpp which looks up the save diretory in the OptionsDB and returns it, similar to GetConfigPath. -Made use of GetSaveDir in various places in code.  Linux save dir creation in InitDirs is still hard-coded to /save within the user dir, however, as I can't test any changes to that code. -Tweaked comments in Directories.h -Replaced boost::filesystem with fs (an alias for the former) in a few places in Directories.cpp -Wrapped some return values in static casts to float to remove compile warnings in Universe.cpp -Added a planet environments map to Species, which maps from PlanetType to PlanetEnvironment, which will allow Species to have customizable environmental preferences. -Modified Species constructor and Dump functions, and added accessor to Species planet environments map. -Added parsing to read environmental preferences from species.txt, and added the relevant text to the species in species.txt. -Tweaked layout and added a make_pair phoenix function to ParserUtil.h and .cpp to allow inserting into associative containers when parsing. -Changed default search path for save dir in OptionsWnd to the user dir.
svn path=/trunk/FreeOrion/; revision=3651

+++ universe/Universe.cpp
@@ -4226,7 +4226,8 @@ void Universe::GenerateEmpires(std::vector<int>& homeworld_planet_ids, const std
                 colors.erase(colors.begin() + color_idx);
             } else {
                 // as a last resort, make up a colour
-                empire_colour = GG::FloatClr(RandZeroToOne(), RandZeroToOne(), RandZeroToOne(), 1.0);
+                empire_colour = GG::FloatClr(static_cast<float>(RandZeroToOne()), static_cast<float>(RandZeroToOne()),
+                                             static_cast<float>(RandZeroToOne()), 1.0f);
             }
         }
